src/main/java/hello/core/discount/RateDiscountPolicy.java
package hello.core.discount;

import hello.core.memeber.Grade;
import hello.core.memeber.Member;
import hello.core.memeber.MemberService;
import hello.core.memeber.MemberServiceImpl;

//정률 할인 정책(할인율 계산)
public class RateDiscountPolicy implements DiscountPolicy{

    private int discountPercent = 10;

    @Override
    public int discount(Member member, int price) {
        if (member.getGrade() == Grade.VIP) {
            return price/discountPercent;
        } else {
            return 0;
        }
    }
}
